Fix git add . on pre-commit hook
"git add ." adds *ALL* changes in the working directory into the commit, whether they had been staged or not. This means the commit would end up with files the user may not have intended to be there. Instead, only files that were already staged should be added, which the -u flag provides.

Signed-off-by: Raúl Santos <4837+borfast@users.noreply.github.com>