fix: fix the numbers for the contributor dependency widget (#262)
We were sending a granularity parameter to TinyBird, which was causing it to return invalid values for the total number of contributors and their percentage of contributions.

We also need to request a higher limit of contributors than the default of 10, because when summing the contribution percentages, we want to get to at least 51%, but sometimes only 10 contributors don't add up to that percentage. By bumping up that limit to 100, it's almost guaranteed we will get to 51%.

There is also another issue with TinyBird when it comes to sorting the results, which sometimes could make the wrong values show in the widget. We grab the total percentage of contributions from the top contributors from the last item in the results, so if those are not correctly sorted, we would get the wrong value.

Signed-off-by: Raúl Santos <4837+borfast@users.noreply.github.com>